Remove checks for versions of Node earlier than 4
