+++ test/integration/connection-pool/yield-support-tests.js
@@ -1,5 +1,28 @@
-var semver = require('semver')
-if (semver.lt(process.version, '1.0.0')) {
-  return console.log('yield is not supported in node <= v0.12')
-}
-require('./yield-support-body')
+var helper = require('./test-helper')
+var co = require('co')
+
+var tid = setTimeout(function() {
+  throw new Error('Tests did not complete in time')
+}, 1000)
+
+co(function * () {
+  var client = yield helper.pg.connect()
+  var res = yield client.query('SELECT $1::text as name', ['foo'])
+  assert.equal(res.rows[0].name, 'foo')
+
+  var threw = false
+  try {
+    yield client.query('SELECT LKDSJDSLKFJ')
+  } catch(e) {
+    threw = true
+  }
+  assert(threw)
+  client.release()
+  helper.pg.end()
+  clearTimeout(tid)
+})
+.catch(function(e) {
+  setImmediate(function() {
+    throw e
+  })
+})
